Jobs that get cleaned up should also be retried
