+++ tests/test_job.py
@@ -1,6 +1,4 @@
 # -*- coding: utf-8 -*-
-from __future__ import (absolute_import, division, print_function,
-                        unicode_literals)
 
 import json
 import time
@@ -12,7 +10,7 @@ from redis import WatchError
 
 from rq.compat import as_text
 from rq.exceptions import NoSuchJobError
-from rq.job import Job, JobStatus, cancel_job, get_current_job, Retry
+from rq.job import Job, JobStatus, cancel_job, get_current_job
 from rq.queue import Queue
 from rq.registry import (DeferredJobRegistry, FailedJobRegistry,
                          FinishedJobRegistry, StartedJobRegistry,
@@ -221,19 +219,6 @@ class TestJob(RQTestCase):
         ts = utcnow()
         job.heartbeat(ts, 0)
         self.assertEqual(job.last_heartbeat, ts)
-
-    def test_persistence_of_retry_data(self):
-        """Retry related data is stored and restored properly"""
-        job = Job.create(func=fixtures.some_calculation)
-        job.retries_left = 3
-        job.retry_intervals = [1, 2, 3]
-        job.save()
-
-        job.retries_left = None
-        job.retry_intervals = None
-        job.refresh()
-        self.assertEqual(job.retries_left, 3)
-        self.assertEqual(job.retry_intervals, [1, 2, 3])
 
     def test_persistence_of_parent_job(self):
         """Storing jobs with parent job, either instance or key."""
@@ -610,7 +595,6 @@ class TestJob(RQTestCase):
 
     def test_job_delete_removes_itself_from_registries(self):
         """job.delete() should remove itself from job registries"""
-        connection = self.testconn
         job = Job.create(func=fixtures.say_hello, status=JobStatus.FAILED,
                          connection=self.testconn, origin='default')
         job.save()
@@ -836,7 +820,6 @@ class TestJob(RQTestCase):
         self.assertListEqual(dependencies, [dependency_job])
 
     def test_fetch_dependencies_returns_empty_if_not_dependent_job(self):
-        queue = Queue(connection=self.testconn)
         dependent_job = Job.create(func=fixtures.say_hello)
 
         dependent_job.register_dependency()
@@ -901,8 +884,6 @@ class TestJob(RQTestCase):
         self.assertFalse(dependencies_finished)
 
     def test_dependencies_finished_returns_true_if_no_dependencies(self):
-        queue = Queue(connection=self.testconn)
-
         dependent_job = Job.create(func=fixtures.say_hello)
         dependent_job.register_dependency()
 
@@ -946,8 +927,6 @@ class TestJob(RQTestCase):
         dependent_job = Job.create(func=fixtures.say_hello)
         dependent_job._dependency_ids = [job.id for job in dependency_jobs]
         dependent_job.register_dependency()
-
-        now = utcnow()
 
         dependencies_finished = dependent_job.dependencies_are_met()
 
@@ -1067,48 +1046,3 @@ class TestJob(RQTestCase):
         self.assertEqual(queue.count, 0)
         self.assertTrue(all(job.is_finished for job in [job_slow_1, job_slow_2, job_A, job_B]))
         self.assertEqual(jobs_completed, ["slow_1:w1", "B:w1", "slow_2:w2", "A"])
-
-    def test_retry(self):
-        """Retry parses `max` and `interval` correctly"""
-        retry = Retry(max=1)
-        self.assertEqual(retry.max, 1)
-        self.assertEqual(retry.intervals, [0])
-        self.assertRaises(ValueError, Retry, max=0)
-
-        retry = Retry(max=2, interval=5)
-        self.assertEqual(retry.max, 2)
-        self.assertEqual(retry.intervals, [5])
-
-        retry = Retry(max=3, interval=[5, 10])
-        self.assertEqual(retry.max, 3)
-        self.assertEqual(retry.intervals, [5, 10])
-
-        # interval can't be negative
-        self.assertRaises(ValueError, Retry, max=1, interval=-5)
-        self.assertRaises(ValueError, Retry, max=1, interval=[1, -5])
-    
-    def test_get_retry_interval(self):
-        """get_retry_interval() returns the right retry interval"""
-        job = Job.create(func=fixtures.say_hello)
-
-        # Handle case where self.retry_intervals is None
-        job.retries_left = 2
-        self.assertEqual(job.get_retry_interval(), 0)
-
-        # Handle the most common case
-        job.retry_intervals = [1, 2]
-        self.assertEqual(job.get_retry_interval(), 1)
-        job.retries_left = 1
-        self.assertEqual(job.get_retry_interval(), 2)
-        
-        # Handle cases where number of retries > length of interval
-        job.retries_left = 4
-        job.retry_intervals = [1, 2, 3]
-        self.assertEqual(job.get_retry_interval(), 1)
-        job.retries_left = 3
-        self.assertEqual(job.get_retry_interval(), 1)
-        job.retries_left = 2
-        self.assertEqual(job.get_retry_interval(), 2)
-        job.retries_left = 1
-        self.assertEqual(job.get_retry_interval(), 3)
-
